# lib/api-zod/src/generated/types/firstAidRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FirstAidRequestSeverity } from "./firstAidRequestSeverity";

export interface FirstAidRequest {
  /** The medical condition or symptom */
  condition: string;
  /** Age of the child (e.g., "6 months", "3 years") */
  childAge: string;
  /** Perceived severity */
  severity: FirstAidRequestSeverity;
  /** Additional context about the situation */
  additionalInfo?: string;
}