# lib/api-zod/src/generated/types/mealPlanRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MealPlanRequestMealType } from "./mealPlanRequestMealType";

export interface MealPlanRequest {
  /** Dietary preferences (vegetarian, vegan, no onion-garlic, etc.) */
  preferences: string[];
  /** Available ingredients */
  inventory: string[];
  /** Number of family members */
  familySize: number;
  /** Nutritional goals (high protein, low oil, etc.) */
  nutritionalGoals?: string[];
  /** Type of meal to plan */
  mealType: MealPlanRequestMealType;
}