# lib/api-zod/src/generated/types/mealPlanResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MealSuggestion } from "./mealSuggestion";

export interface MealPlanResponse {
  meals: MealSuggestion[];
  nutritionSummary?: string;
  tips?: string;
}