# lib/api-zod/src/generated/types/mealSuggestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MealSuggestion {
  name: string;
  nameHindi?: string;
  description: string;
  ingredients: string[];
  prepTime: string;
  nutritionHighlights?: string;
  servings?: number;
}